Renaming and removing unuused components
