+++ src/views/CollectionBlock.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useContext, useEffect, useState } from 'react';
 import styled, { keyframes } from 'styled-components';
-import { BlockHeader, NFTTokens } from 'components';
+import { BlockHeader, Tokens } from 'components';
 import { AppStateContext, IConnectData } from '../AppContextWrapper';
 import { ICollection, IToken } from 'SDK/ContractsSDK';
 import { useParams } from 'react-router-dom';
@@ -37,7 +37,7 @@ const CollectionBlock: FC = () => {
         creator={collection.creator}
         description={collection.description}
       />
-      <NFTTokens tokens={tokens} />
+      <Tokens tokens={tokens} />
     </CollectionBlockWrapper>
   );
 };
