+++ src/views/MyCollections.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useCallback, useContext, useEffect, useState } from 'react';
 import styled, { keyframes } from 'styled-components';
-import { BlockHeader, Collections, NFTTokens } from 'components';
+import { BlockHeader, Collections, Tokens } from 'components';
 import { ICollection, ICreator, IToken } from 'SDK/ContractsSDK';
 import { AppStateContext, IConnectData } from '../AppContextWrapper';
 import { useParams } from 'react-router-dom';
@@ -60,7 +60,7 @@ const MyCollections: FC = () => {
         </CollectionsTab>
       </ActiveContent>
 
-      {activeTab === "Tokens" ? <NFTTokens tokens={tokens}/> : <Collections collections={collections}/>}
+      {activeTab === "Tokens" ? <Tokens tokens={tokens}/> : <Collections collections={collections}/>}
     </MyCollectionsWrapper>
   )
 };
